Lots of improvements to the rents_basic and rents_advanced work flow, but note that I am experimenting with a raw sql query for rents_basic as an alternative to our several sqlalchemy queries, none of which I can get to work fully. I am sure one of us may be able to overcome our joint disability on this front, but, at the moment, my raw sql query works much faster and much better than any ORM alternative I have tried.
I have introduced a global function next_rent_date() which can be
called in any html template.  It is a pity it takes four parameters,
but it is used all over mjinn and I am confident it is the right way to
go.

I have dropped date_m from models.py and from my db and suggest you drop
the table from your db.

I am not sure where J is up to with HR statuses, as I think line 25 in
views.headrent.py should be removed if this value is available globally.

I think we may be able  get rid of all combodicts in main.common and
also, hopefully, get__rents_fdict, if my work in main.rent_filter comes
to fruition.